・警告が出ていた部分を一部修正 ・プレイヤーの走り方やアイテムの使用方法を表すアイコンを追加 ・アイテムの使用方法を右クリックに変更

+++ GameProgramming/CMatrix.cpp
@@ -55,7 +55,7 @@ CMatrix CMatrix::Scale(float sx, float sy, float sz)
 //RotateY(角度)
 CMatrix CMatrix::RotateY(float degree) {
 	//角度からラジアンを求める
-	float rad = degree / 180.0f * M_PI;
+	float rad = degree / 180.0f * (float)M_PI;
 	//単位行列にする
 	Identity();
 	//Y軸で回転する行列の設定
@@ -71,7 +71,7 @@ CMatrix CMatrix::RotateY(float degree) {
 CMatrix CMatrix::RotateZ(float degree)
 {
 	//角度からラジアンを求める
-	float rad = degree / 180.0f * M_PI;
+	float rad = degree / 180.0f * (float)M_PI;
 	//単位行列にする
 	Identity();
 	//Y軸で回転する行列の設定
@@ -86,7 +86,7 @@ CMatrix CMatrix::RotateZ(float degree)
 CMatrix CMatrix::RotateX(float degree)
 {
 	//角度からラジアンを求める
-	float rad = degree / 180.0f * M_PI;
+	float rad = degree / 180.0f * (float)M_PI;
 	//単位行列にする
 	Identity();
 	//Y軸で回転する行列の設定
